refactor: rename dir where app lives in to internal
