+++ internal/controller/redirect/input/find_test.go
