+++ internal/database/output/document/read.go
