+++ internal/queue/asynq/server_adapter_test.go
@@ -5,7 +5,7 @@ import (
 	"errors"
 	"testing"
 
-	"github.com/dorianneto/url-shortener/src/job/jobtest"
+	"github.com/dorianneto/url-shortener/internal/job/jobtest"
 	"github.com/hibiken/asynq"
 )
